Flatten directory structure; Move utils to files they are used
