+++ examples/segmentation/fastsam-multistream/src/cpp/image_encoder.h
